+++ promiumbookmarks/plugins/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from __future__ import print_function
+
+class PromiumPlugin(object):
+    def process_bookmarks(self, bookmarks_obj):
+        return bookmarks_obj
+
+